Add files via upload
This is a fun little project that I like to do in my spare time. It is a number-guessing game written in Python. The basic idea is to have the computer produce a random number between 1 and 100 and then have the user try to guess it. If the user guesses correctly, they win! If not, the computer tells them whether their guess was too high or too low, and they get another chance. It is a simple concept but a lot of fun to play.